requirements.txt
flask==2.0.1
Pillow==8.3.1
python-dotenv==0.19.0
flask-cors==3.0.10